fix a leak of platform information via navigator.appVersion.
The `navigator.appVersion` property returns a string that can
identify a user's browser and platform. It is documented [on mdn](https://developer.mozilla.org/en-US/docs/Web/API/NavigatorID/appVersion).

I noticed on the Brave browser running on Linux, it looked to be
a copy of the User-Agent string. I decided to update client.js
to return the UA string from `navigator.appVersion` - same as
we are doing for `navigator.userAgent`.